split user secrets into separate store
