+++ openhands/server/user_auth/user_auth.py
@@ -8,6 +8,7 @@ from pydantic import SecretStr
 from openhands.integrations.provider import PROVIDER_TOKEN_TYPE
 from openhands.server.settings import Settings
 from openhands.server.shared import server_config
+from openhands.storage.data_models.user_secrets import UserSecrets
 from openhands.storage.settings.settings_store import SettingsStore
 from openhands.utils.import_utils import get_impl
 
@@ -44,6 +45,10 @@ class UserAuth(ABC):
         settings = await settings_store.load()
         self._settings = settings
         return settings
+    
+    @abstractmethod
+    async def get_user_secret_store(self) -> UserSecrets | None:
+        """Get the user's secret store"""
 
     @classmethod
     @abstractmethod
