+++ openhands/storage/data_models/settings.py
@@ -12,7 +12,7 @@ from pydantic.json import pydantic_encoder
 
 from openhands.core.config.llm_config import LLMConfig
 from openhands.core.config.utils import load_app_config
-from openhands.integrations.provider import SecretStore
+from openhands.storage.data_models.user_secrets import UserSecrets
 
 
 class Settings(BaseModel):
@@ -29,7 +29,7 @@ class Settings(BaseModel):
     llm_api_key: SecretStr | None = None
     llm_base_url: str | None = None
     remote_runtime_resource_factor: int | None = None
-    secrets_store: SecretStore = Field(default_factory=SecretStore, frozen=True)
+    secrets_store: UserSecrets = Field(default_factory=UserSecrets, frozen=True)
     enable_default_condenser: bool = True
     enable_sound_notifications: bool = False
     user_consents_to_analytics: bool | None = None
@@ -66,10 +66,10 @@ class Settings(BaseModel):
         custom_secrets = secrets_store.get('custom_secrets')
         tokens = secrets_store.get('provider_tokens')
 
-        secret_store = SecretStore(provider_tokens={}, custom_secrets={})
+        secret_store = UserSecrets(provider_tokens={}, custom_secrets={})
 
         if isinstance(tokens, dict):
-            converted_store = SecretStore(provider_tokens=tokens)
+            converted_store = UserSecrets(provider_tokens=tokens)
             secret_store = secret_store.model_copy(
                 update={'provider_tokens': converted_store.provider_tokens}
             )
@@ -77,7 +77,7 @@ class Settings(BaseModel):
             secret_store.model_copy(update={'provider_tokens': tokens})
 
         if isinstance(custom_secrets, dict):
-            converted_store = SecretStore(custom_secrets=custom_secrets)
+            converted_store = UserSecrets(custom_secrets=custom_secrets)
             secret_store = secret_store.model_copy(
                 update={'custom_secrets': converted_store.custom_secrets}
             )
@@ -89,7 +89,7 @@ class Settings(BaseModel):
         return data
 
     @field_serializer('secrets_store')
-    def secrets_store_serializer(self, secrets: SecretStore, info: SerializationInfo):
+    def secrets_store_serializer(self, secrets: UserSecrets, info: SerializationInfo):
         """Custom serializer for secrets store."""
         return {
             'provider_tokens': secrets.provider_tokens_serializer(
